Addressed some of Sahil comments

+++ Autopilot/PathManager/Inc/cruisingMode/cruisingModeStages.hpp
@@ -1,6 +1,8 @@
 #ifndef CRUISING_MODE_STAGES_HPP
 #define CRUISING_MODE_STAGES_HPP
 
+#include <cstdint>
+
 #include "cruisingModeStageManager.hpp"
 #include "cruisingMode.hpp"
 
@@ -22,7 +24,7 @@ class CruisingFlight : public CruisingModeStageManager {
         CruisingFlight& operator=(const CruisingFlight& other);
 
         WaypointManager cruising_state_manager;
-        int waypoint_id_array[PATH_BUFFER_SIZE]; // Stores ids of the waypoints in the flight path in the order that they are executed
+        uint16_t waypoint_id_array[PATH_BUFFER_SIZE]; // Stores ids of the waypoints in the flight path in the order that they are executed
         static _WaypointManager_Data_In _input_data;
         static _WaypointManager_Data_Out _output_data;
         static _CruisingState_Telemetry_Return _return_to_ground;
